requirements.txt
python-telegram-bot==8.1.1
pymysql
icalendar
git+git://github.com/irgangla/icalevents@master#egg=icalevents
requests_html
googlemaps
python-dateutil
